+++ src/main/java/com/example/kvjp/repository/ApartmentRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ApartmentRepository extends JpaRepository<Apartment, Long> {
+public interface ApartmentRepository extends JpaRepository<Apartment, Integer> {
     Apartment findByName(String name);
 
     @Query(value = "select a from Apartment a " +
