+++ src/main/java/com/example/kvjp/repository/LeasesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LeasesRepository extends JpaRepository<Leases, Long> {
+public interface LeasesRepository extends JpaRepository<Leases, Integer> {
 }
